Update patch set 6
Patch Set 6: Code-Review+1

I am okay with the change, but not sure why this service account is needed and what are the implications of adding it in the policy charts. Not much description in the associated jira as well.
Do we have any documents/wiki referring to the work?

Patch-set: 6
Reviewer: Gerrit User 3239 <3239@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1