Update patch set 7
Patch Set 7: Code-Review+2

> Patch Set 6: Code-Review+1
> 
> I am okay with the change, but not sure why this service account is needed and what are the implications of adding it in the policy charts. Not much description in the associated jira as well.
> Do we have any documents/wiki referring to the work?

Hi,
There are several reasons for this change:
If service account is not set, the default one with too many authorization will be picked. 
So it’s a best practice for security to add a service account to all deployments/statefulsets. 
Second reason is that this service account may be used by other Kubernetes services and in particular service mesh. 
With service mesh, this service account may be used to enforce some authorization (for example, we could create a rule stating that only clamp fe can talk to clamp be if it makes sense). 
So that’s why we’re adding it although it’s perfectly harmless for policy for now.

Patch-set: 7
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2