Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> Hi Bartek,
> log is a deprecated component so I don't know what's the policy for update.
> Also there's this licensing issue that needs to be tackled at SECCOM / TSC level I guess

On last DDF there was a presentation on ONAP logging and future plans for the project were presented, it wasn't communicated in any way that it's deprecated and not accepting any changes. Looking at the OOM repo log I can see some more or less recent changes to it. For this reason I also asked on OOM Slack chanel what is the situation with LOG but got no input whatsoever.

Looking at the page you referenced it seems that licensing only concerns Elasticsearch and Kibana, Logstash as I understand it is not affected?

Patch-set: 1