Update patch set 1
Patch Set 1: Code-Review-2

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > (1 comment)
> > 
> > Hi Bartek,
> > log is a deprecated component so I don't know what's the policy for update.
> > Also there's this licensing issue that needs to be tackled at SECCOM / TSC level I guess
> 
> On last DDF there was a presentation on ONAP logging and future plans for the project were presented, it wasn't communicated in any way that it's deprecated and not accepting any changes. Looking at the OOM repo log I can see some more or less recent changes to it. For this reason I also asked on OOM Slack chanel what is the situation with LOG but got no input whatsoever.
> 
> Looking at the page you referenced it seems that licensing only concerns Elasticsearch and Kibana, Logstash as I understand it is not affected?

Sorry I'm here with Sylvain. According to:
https://www.elastic.co/guide/en/logstash/current/installing-logstash.html

the default package includes non-free source code. This is not compatible with our Apache License. If you want to touch to anything you need to use -oss image as a bare minimum.

Even if we do this, the talk that was presented on logging is rather about plans not the current state. Current logging project has been stalled and obsoleted by the TSC quite some time ago but we continue using it as currently we don't have a ready alternative to it. If you would like to make changes to the logging please provide a really comprehensive justification for it

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2