Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> that's hackish in my opinion. I'd prefer so that you pass explicitly --variables to env subst instead to make it more explicit and more obvious for the reader

Here the case is more complex.
We have a configuration property, where part of the path contains the $variables.

The envsubst gets a number of envs added to the container runtime and the idea is, that envsubst uses the env values to replace the $variables present in a configuration file (stored in a configMap).

PRH is probably the only component currently using the "$variables" in the configuration property, used as a path.

If we updated the envsubst container with envs corresponding to there parameters, the values for these envs would be really sth like: customer="$customer", tenant="$tenant".

Not sure, how that will really behave, but it could be an option overall too.
Your view?

Patch-set: 2
CC: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>