Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> that's hackish in my opinion. I'd prefer so that you pass explicitly --variables to env subst instead to make it more explicit and more obvious for the reader

Hi Krzysztof - As Damian noted, the application (PRH) does the substitution based on incoming VES event data (hence actual values are not known at deployment time). As the pattern used for in-app substitution was same as envsubst container, the configurations were getting replaced before being made available to PRH app.

This is an one-off case; we are exploring other option to replace PRH in-app substitution via different pattern. 

Will be good if this can approved as this is blocking intergration smoke case (SO-3708 & SO-3710)

Patch-set: 2