Update patch set 5
Patch Set 5: Code-Review+1

Looks good to me as a temporary solution (and I agree with Vijay).

The DCAE-SDK 1.8.6 (PRH uses currently 1.6.0) will resolve this issue permanently.
There is a plan to update the PRH code to change the default definition of variables in a path, and to migrate to that newest DCAE-SDK, but let`s fix the smoke tests first with this change.
Then we can really see after that DCAE-SDK update (1.6.0 -> 1.8.6), that we haven`t messed that thing up again.

Patch-set: 5
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1