Update patch set 7
Patch Set 7: Code-Review+1

> Patch Set 7: Code-Review-2
> 
> Hi Guys,
> 
> I trigger the gating on this one once again as it failed but before we proceed with this change could you please:
> 
> 1) Clearly describe why this version bump is not applicable to master
> 
> or
> 
> 2) Submit the patch to master first and we will merge it after that one gets merged

We take the approach that changes that are only destined for Honolulu (or Honolulu Maintenance) should only be made on the honolulu branch.  There is no benefit in pushing it to master first.  That's the approach we've taken with the Policy repos and see no reason to treat OOM any differently.  With OOM, in particular, pushing to master first would be a serious impediment, given how long it takes to get reviews merged.

That being said, we will likely push a review for the master branch next week, but that will contain a first cut at images for Istanbul, all of which will have different version numbers than they do in honolulu.  (The property files used within the pods will likely be different, too.)

Patch-set: 7
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1