Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-2
> 
> I haven't got answer to my previous question
> 
> Am I missing sth or you are exposing new http NodePorts with this change?
> If yes then what's the SECCOM opinion on this?

Hi Krzysztof, Maybe I misunderstood, I though you wanted me to add 'plain_port' to mandate the use of http(according to the documentation). I am able to access the service inside the kubernetes cluster using  port number 1680, and outside the cluster using node port 30408.
Here is the blueprint which I had referred for helm transformation.
https://gerrit.onap.org/r/c/dcaegen2/platform/blueprints/+/120607

For the second question yes it is a NodePort as mentioned in the blueprints.
SECCOM opinion for exposing NodePort? @Vijay or @Krzysztof could you please explain? I wasn't aware I had check SECCOM for this since it was already a NodePort in the cloudify blueprint.

Patch-set: 5