Update patch set 13
Patch Set 13:

@Sylvain

I have heard that you have some doubts about manual upgrading CRD.
Unfortunately it looks that helm 3 doesn't support upgrade CRD in any way.

In the official documentation we can find that: "These CRDs are not templated, but will be installed by default when running a helm install for the chart. If the CRD already exists, it will be skipped with a warning." (https://helm.sh/docs/chart_best_practices/custom_resource_definitions/)

So, this is independent of the version we specified in CRD (v1 or v2) - "helm install" detects that CRD exists and skips installing the new one. Helm 3 also doesn't manage CRD's lifecycle, so it doesn't delete it when uninstalling the chart.

It is worth mentioning that Kubernetes correctly handles the new CRD after deploying the old one. It is sufficient to execute kubectl apply, without removing the old CRD first. It confirms that adding a new optional field (updateEndpoint) doesn't break CMPv2Issuer CRD API and the change is backwards compatible, so it isn't necessary to change CRD version from v1 to v2.

Patch-set: 13