Update patch set 2
Patch Set 2:

and those haven't changed in a while so not sure why it should have stopped working all of a sudden.  I noticed in yesterday's daily gating that one of the policy healtchecks timed out, but it was working earlier this week.  Not sure if that's what happened here - where can we find the results?

Patch-set: 2
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>