[SO] Check for job instead of container for so-mariadb-config
so-bpmn-infra deployment init container waits on
so-mariadb-config pod to complete instead of checking job
status itself. This causes issues when the completed config
job container finishes and gets cleaned up due to garbage
collector, after which so-bpmn-infra pod gets stuck in init
state forever.

Fixing other instances of waiting for so-mariadb-config job
as well.

Issue-ID: OOM-2799
Signed-off-by: Prabhjot Singh Sethi <prabhjot@aarnanetworks.com>
Change-Id: Ib6d1b53d52c178386a743558c4b45b8f8748487b