Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> (1 comment)
> 
> Let's talk about it during our OOM meeting. Maybe we would need to "revise" a little bit our readiness image because adding this script to so many components may be unmaintainable

The readiness image will first check if a container (given as an argument) is part of a Job, deployment, statefulset etc. If it is part of a Job then it will wait for the Job status to go to "Complete" which will only occur once all containers have finished. Since we need our new container to wait until the initial Job container is finished and then send our quitquitquit curl command this will not work. So this script I have added has small changes from readiness script, that will check the container status, as opposed to the Job status.

Happy to discuss it in OOM meeting, I agree it could become unmaintainable. Possibly adding a single script file to onboard to each new container, or editting the "oom/readiness" repo itself and adding a new route through the code for this scenario.

Patch-set: 10