Update patch set 2
Patch Set 2:

Hi Dan,

As we discussed yesterday this change should be squashed with the SO change above into a single commit.
Would you mind to do it so that I can merge this?
SO change just got a fully green gating to after this "technical" change we should be good to go

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>