Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> I've noticed that SDC also has troubles registering as a producer for the distribution topic:
> "SDC-DISTR-NOTIF-TOPIC-AUTO" STATUS = "CREATED"	PUT http://message-router.onap:3904/topics/SDC-DISTR-NOTIF-TOPIC-AUTO/producers/iPIxkpAMI8qTcQj8 will send credentials over a clear channel.
> 2021-08-25T14:40:57.390Z	[pool-17-thread-1]	INFO 	c.att.nsa.apiClient.http.HttpClient	AuditMessage=ACTION = "CreateDistributionTopic" D_ENV = "AUTO" TOPIC_NAME = "SDC-DISTR-NOTIF-TOPIC-AUTO" STATUS = "CREATED"	PUT http://message-router.onap:3904/topics/SDC-DISTR-NOTIF-TOPIC-AUTO/producers/iPIxkpAMI8qTcQj8 (as iPIxkpAMI8qTcQj8) ...
> 2021-08-25T14:40:57.455Z	[pool-17-thread-1]	INFO 	c.att.nsa.apiClient.http.HttpClient	AuditMessage=ACTION = "CreateDistributionTopic" D_ENV = "AUTO" TOPIC_NAME = "SDC-DISTR-NOTIF-TOPIC-AUTO" STATUS = "CREATED"	\t--> HTTP/1.1 500 Server Error

Hi @Christophe,
Trying to debug this and not having much success.
Which pod in SDC does the creation of this topic?
I am not familiar with the apikeys setup in MR and I see the failing components CDS, SDCN etc are using the cambria client, which from what I can see is just another implementation of the dmaapClient.
Can anyone on SDC give some guidance here as I have no knowledge, or access to the CambriaClient code base.

Patch-set: 6