Update patch set 6
Patch Set 6:

Hi @Krzysztof,
This seems to be a timing issue. Recreated it locally with just enough for cds to deploy.
Then tore it down, redeployed and the cds-sdc-listener came up ok.

Is there a way to force DMaaP to be deployed earlier in the sequence? Maybe directly after AAF.

Patch-set: 6