Update patch set 6
Patch Set 6:

Hi Guys,
What I understand is now that dmaap starts has changed a little bit (it's slower?) and when SDC tries to create a topic, it has an error.
there are several ways to fix it (one being proposal from fiachra to "force" onap to deploy DMaaP sooner).
But, IMHO, I think it's not the right way as when we'll have a "fast" SDC start and a "fast" deployment of ONAP, we'll face the issue again.
So I think we should more to the work in SDC:

* if having this topic is mandatory for SDC, then it should crash if not having it
* if having this topic is optional, it should retries once in a while to check if the topic creation is doable

@SDC team, what's your point of view?

Patch-set: 6
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>