Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Hi Fiachra,
> 
> so what's the result of discussion with Christoph?

Hi Krzysztof,
Not good news I'm afraid. SDC (and maybe others) are using pre loaded zookeeper apiKeys.
On upgrading the kafka/zk images, I removed what I thought was redundant data, but in fact it was preloading the keys.
To remove this problem, all MR clients that are using these keys would need to change also. 
So... my proposal is to replace the keys and look to flag this with the relevant parties. (SDC and possibly Policy)

Patch-set: 6