Update patch set 3
Patch Set 3:

patchset 3 picks up the recently merged change that adds a chart for dcae-snmptrap-collector and modifies those charts to add the Consul key delete job.
dcae-snmptrap-collector is not deployed by default in a DCAE Helm installation.  It's intended to be
deployed on demand after the initial deployment of ONAP and DCAE.  I verified that on-demand deployment and undeployment (via helm install and helm delete) works in my test environment.  The
Consul delete job runs correctly during undeployment.

Patch-set: 3