Update patch set 8
Patch Set 8: Code-Review-1

SO/AAI issue due to model-version not found

551743a3-37de-410b-bc4c-142654d83b2b"},"requestStatus":{"requestState":"FAILED","statusMessage":"STATUS: Error Source: UNKNOWN, Error Message: Exception in org.onap.so.bpmn.infrastructure.aai.tasks.AAICreateTasks.createServiceInstance Error calling A&AI. Request-Id=a9b31c02-a2ba-4731-9d67-068ac455c53c Resource not found for PUT using id business/customers/customer/basicvm-customer/service-subscriptions/service-subscription/basic_vm/service-instances/service-instance/6876a72f-9adf-4caf-83f1-4fdb98892050 (msg=Node Not Found:object located at service-design-and-creation/models/model/96d84d26-f3c2-45df-8204-1c1a54a400c5/model-vers/model-ver/0500d784-8d0b-49ff-ad97-e77a33e28645#model-version not found) (ec=ERR.5.4.6114) FLOW STATUS: All Rollback flows have completed successfully ROLLBACK STATUS: Rollback has been completed successfully.","percentProgress":100,"timestamp":"Fri, 17 Sep 2021 14:48:09 GMT"}}}

In the onapsdk we use an "old version" of the API (v20)
https://gitlab.com/Orange-OpenSource/lfn/onap/python-onapsdk/-/blob/develop/src/onapsdk/configuration/global_settings.py
the onboarding is still fine using this version of the API
but the instantiation is failing

does it mean that the v24 breaks the API compatibility? and moving the tester to v24 would fix the issue? We could give a try but if so it means anyway the the v20 would not be compatible anymore.

Patch-set: 8
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1