Update patch set 3
Patch Set 3:

I believe the /sdc/v1/catalog/resources?resourceType=VF API is good. I suspect the reason it is returning no VFs is due to an issue in importing the VSPs. We have identified an issue in the import and will push a new patchset with dockers where that issue is addressed.

Patch-set: 3