Update patch set 5
Patch Set 5:

The test cases that are failing are failing in SO rather than SDC, its difficult to see a connection to the changes in SDC. Looking at gating results from other reviews, the same failures can be seen, e.g. : https://logs.onap.org/onap-integration/gating//119248-16/index.html, and they passed on a re-run of the tests with no changes. 
Are these failures just down to instability in the tests?

Patch-set: 5