Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Oups...Sorry for the previous comment.
> sphinxcontrib.plantuml is in the requirements-docs.txt
> Anyway, we should wait for new lfdocs-conf version to avoid to include funcparserlib here

this is good news.
OOM gate is currently broken because of that.
The changes in this relation intend to reestablish a correct jenkins verified +1 voting on the gate to ease reviews.
Until lf-docsconf is fixed, I think it is preferrable to merge this change to fix oom gate and proposes a new change as soon as the new lf-docsconf package is out.

Patch-set: 3
Reviewer: Gerrit User 3685 <3685@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>