Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > OK I think i know, something changed (may be on AAF ?)
> > The default login/password demo/demo123456! does not work anymore.
> > 
> > It must be demo@people.osaaf.org/demo123456!
> > 
> > That could explain the 401 😞
> 
> 
> Hi Seb, Our helm charts are using the credentials aafDeployFqi: deployer@people.osaaf.org
>   aafDeployPass: demo123456! for certInitializer. This needs to be updated for the basic clamp test to pass?

No this is not related to your fix, I think the problem comes from an ancestor PR where the version 6.1.2 is introduced for Clamp BE/FE. During the testing we have noticed that the demo user was not working anymore, it must be demo@people.osaaf.org (may be due to Cadi changes or something like that ??). Anyway it requires a fix in the Clamp tests, the user must be changed from "demo" to "demo@people.osaaf.org".
Unless i'm mistaken, I think that should solve the issue.

Patch-set: 7
Reviewer: Gerrit User 23 <23@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>