[CCSDK] Workaround for naming service host name verification issue
The naming service query to policy manager is failing host name
verification.  While this issue is under investigation, it can be
worked around by configuring the naming service to disable host
name verification.  This requires a new version of the network-name-gen
microservice (1.2.1), and a change to set a new env variable to
disable host name verification (enabled by default if this variable
is unset).

Issue-ID: CCSDK-3501
Signed-off-by: Dan Timoney <dtimoney@att.com>
Change-Id: Ia471cd27ce16b6e79a3ce6708e08c7d5f239feb3