Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > Hi. I know this is unrelated but what is this:
> > > "Using generated security password: 96497452-1ea5-4b06-a34e-4b96467f0b4b" ?
> > > 
> > > What is security password used for? Why is it generated in the container not within helm chart? Why it is outputed to logs?
> > 
> > Hi Krzysztof. Can I ask you where you are seeing this? I cannot see it in the logs of the Jenkins Jobs. Perhaps I missed it?
> 
> https://logs.onap.org/onap-integration/gating//125680-1/k8s/k8s/kubernetes-status/pod-onap-policy-clamp-cl-pf-ppnt-b58fb6775-7f2ff-policy-clamp-cl-pf-ppnt-logs.html

Ah yes, I see. This happens because including the spring-boot-starter-security dependency in your Spring Boot project will cause it to auto-generate the password if it is not specified in the application.properties or application.yaml. It will randomly generate a password and print it to the logs. This password is not actually used in the application at all. There is a fix on the way through got this also here: https://gerrit.onap.org/r/c/policy/clamp/+/125610 and here: https://gerrit.nordix.org/c/onap/oom/+/11285

Patch-set: 1