Update patch set 1
Patch Set 1:

> Patch Set 1: -Code-Review
> 
> > Patch Set 1: Code-Review+1
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > please assign correct Issue-ID
> > 
> > In general we don't use Issue-ID in OOM. We place them only to make gerrit auto-checks happy
> 
> hard to agree with this not 'good practice', and those changes just 'for change', and instead of fixing/solving outdated filebeat-images with unsupported Python 2, those commits disabling creation of logs and nobody knows reason or 'why', and what is the plan, are you going to replace filebeat with something?, are you going to upgrade to latest version?

I understand your concerns. Together with Sylvain we just find it super difficult and fragile to use Jira to document our features. We tend to prefer good commit messages and make sure that the documentation is as close to the code as possible (either directly in comment to template or in RTD).

Nothing is getting disabled in those commit. Just the ability to disable all filebeat sidecars with one option is added but the default behavior (as is now) is preserved.

We would love to upgrade this but there is a license issue in newer filebeat versions. Those new containers includes some non-open source components which would be incompatible with the ONAP license policy

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>