[COMMON] Fix shopt nullglob bashism
pointed out by checkbashisms.

shopt, and therefore all the options it provides (only nullglob here)
are not defined by POSIX and are bash-specific.

The specific option nullglob is well documented in the bash manual.
In short, it makes filename globbing patterns that do not match any
filenames simply expanded to nothing rather than remaining unexpanded.

$ echo my*file
my*file
$ shopt -s nullglob
$ echo my*file

This is actually useless in the script where it is used here
since no blob is used to expand such filenames.

Issue-ID: OOM-2643
Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: Id336c17a46c6e5c0748562e7cd8f0efa5570ce6d