Update patch set 8
Patch Set 8:

(10 comments)

Hello

my feedback in the text below.

Hope this helps
Guillaume

> Patch Set 8:
> 
> (7 comments)
> 
> Hello Guillaume,
> 
> I see that we have 3 main disagreements:
> 
> * a single "big" patch. Most of it is the add of gitlab files (ci, markdown, ...) so I think it's a real "single entity" and it cannot be cut down as we'll use either gitlab or jenkins. half a gitlab is not possible.
> 

I think that changes 126205 126240 126207 and 126199 have demonstrated the absolute opposite i.e. this change can perfectly be split. I can see apart the Gitlab migration itself at least these actions:
  - checkbashims commands move to a dedicated script
  - clean-up of some obsolete code
  - a pylintrc creation and the enforcement of pylint
  - the split of the linters job in tox into several git section (though this point is discutable since this is also the Jenkins situation)
  - the replacement of some pre-commit repo by shell scripts in CI

If you think that there is good reason why any of this point is somehow linked to the GitLab migration itself, my opinion ist that this should be more developped in the commit message.


> * not use of tox in the CI. As already said, tox is a wrapper that can makes sense when you're using shared systems (such as jenkins shared runners). In other CI systems (travis ci, gitlab ci, ...), the isolation is done at job level and then you don't need another wrapper. 
>   It doesn't mean that we remove tox.ini but it means that we won't use it in CI. The consequence is that we need to be sure that tox.ini files is aligned with ci file (gitlab-ci.yml here) and that's what we'll do

I think the term wrapper is a bit reductive to what tox brings in term of isolation and dependencies management. But honestly, I am less worried here by the absence of tox than by the regression brought by this change in term of (both CI and OS) portability and maintenability, especially configuration  and code or functions duplications .

> 
> * use of 2 custom scripts instead of precommit script for tabs and trailing whitespaces detection. Here I think that creating a Dockerfile just for these 2 functions (38 and 39 lines in bash, license included and 37/102 lines in Python) doesn't worth the efforts.

If so, why not running all pre-commit and tox jobs in the same and only container ?
These would reduce drastically the number of docker images proposed here.
The configuration in  https://gerrit.onap.org/r/c/oom/+/126207/1 took only 6 minutes 9 seconds to perform all tox and pre-commit linter actions in Gitlab-ci.

>   You can argue that the shell scripts supports only super basic whitelist but today there  are no need on trailing whitespaces of such whitelists and the one for tabs is only on Makefile (the other part of your whitelist are not existent in OOM repository)

Agreed this is details (but this does not give any guarantee the rest of the whitelist won't be found in the future).

>   Again, precommit will not disappear and anyone wanting to use it (with or without tox) will be able to use it.
>   But it won't be used in the CI.
>   My plan is to reassess every couple of monthes if we have a deviation between what has been found by the 2 ci scripts and pre commit. If we have such deviation, then we'll move to precommit also in CI.

My feeling is that this is not completely coherent.
From a maintanibility and troubleshooting standpoint, if precommit or tox are really so unadapted to the CI, I would prefer they simply disappear from the codebase and that we keep only one solution so that we avoid code/configuration/functions duplications and so that contributors directly use the solution enforced in the CI.

> 
> 
> I do acknowledge that you don't agree with this patch but the OOM committers have decided to move on. Therefore, I'll submit this patch as soon as Krzysztof puts back his +2.

I invite you to read the following link in OpenStack wiki.
https://wiki.openstack.org/wiki/GitCommitMessages
It sums up pretty well some problems I see here and it well explains the importance to segment one's work in several atomic commit/pieces to ease the review and to identify potential flaws or regressions among other things.
As it is, I think this change doesn't meet the criteria of the OpenStack community to be merged at least because it does not describe enough in its commit message the (non-trivial) actions performed and the reason why.

Patch-set: 8
Reviewer: Gerrit User 3685 <3685@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>