Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > (10 comments)
> > 
> > Hello
> > 
> > my feedback in the text below.
> > 
> > Hope this helps
> > Guillaume
> > 
> > > Patch Set 8:
> > > 
> > > (7 comments)
> > > 
> > > Hello Guillaume,
> > > 
> > > I see that we have 3 main disagreements:
> > > 
> > > * a single "big" patch. Most of it is the add of gitlab files (ci, markdown, ...) so I think it's a real "single entity" and it cannot be cut down as we'll use either gitlab or jenkins. half a gitlab is not possible.
> > > 
> > 
> > I think that changes 126205 126240 126207 and 126199 have demonstrated the absolute opposite i.e. this change can perfectly be split. I can see apart the Gitlab migration itself at least these actions:
> >   - checkbashims commands move to a dedicated script
> >   - clean-up of some obsolete code
> >   - a pylintrc creation and the enforcement of pylint
> >   - the split of the linters job in tox into several git section (though this point is discutable since this is also the Jenkins situation)
> >   - the replacement of some pre-commit repo by shell scripts in CI
> > 
> > If you think that there is good reason why any of this point is somehow linked to the GitLab migration itself, my opinion ist that this should be more developped in the commit message.
> 
> I share your love to small incremental changes believe me but taken into account that we cannot have our CI both in jenkins and in gitlab in the same time I see no value in adding small commits that cannot be tested one by one compared to 1 commit and instant switch to gitlab for CI
> 
> > 
[...]
> 
> I agree that the commit message is not best-in-class standards of various open source communities. Normally I'd probably ask Sylvain to refine that but provided that I'm available only until 23rd of December and then I'll need to step down for 6 months I don't want to block Sylvain with this transition to gitlab as with only a PTL this process can be super painful for him.

Thanks Krzysztof
I appreciate your true and sincere feedback.
And I share your concern about the difficulty to maintain alone a big project such as OOM.
Even 2 people is a too few IMO.
I agree dealing with such a migration is quite stressfull and painful in these conditions.
One week is quite fast, especially if a pipeline run last 6 hours...

I don't want to block you with that. But if you really need some support,
I am actually  convinced this change cannot be reviewed by anyone external to the project 
in one big block as it is pushed here. 
As a GitLab newbie I did not manage to make it work in my sandbox out-of-the-box.
Maybe some configuration options in your GitLab instances are missing.
If so, some guidance would have been helpful.

Apart the divergence on the orientation taken with tox,
I already detected a few issues in this review but there are probably more.
So I had no other choice to split this change to evaluate how those new profiles behave in my Gitlab sandbox.
I did actually succeed to make most of them work without going through the whole migration.
So I disagree with the impossibility to deal with such a problematic step by step.

Anyway, as a result, it seems some profiles were not really enforced and some entrypoint had to be overriden to make this configuration work.
I'll push you my whole rework of this change with the topic "tox-removal-gitlab-migration-split".
(I share Cédric's comments and point of view on tox defacto removal. Using tox as a mitigation layer should be encouraged so that external contributors can reproduce the CI behavior locally.)
I propose you have a look at it.
Feel free to comment it and to reuse whatever material it brings at your convenience if you may encounter the same issues.

Hope this helps

I wish you a merry Christmas to you and your family.
Guillaume

Patch-set: 8