[PMS] Fix of configuration update
The application.configuration.json can be updated and read by using a
PMS REST API. Currently the file is mapped to a configmap. A file
mapped to a configfmap can not be updated.

The fix moved the file to
/var/policy-management-service/application_configuration.json
(which is a persisten volume).
The possibility to read/update the application_configuration.json via
a config map is removed.

Issue-ID: CCSDK-3560
Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>
Change-Id: I5e8080c80297f36007ff0cea58b1ab30b4e6510f