Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4: Code-Review-1
> > 
> > doesn't seems to work: https://logs.onap.org/onap-integration/gating//127471-4/full/full/report.html#tags?health-holmes
> 
> Hi Sylvain,
> Is the report based on the patch I submitted? If the patch has not been merged, how could it be verified? 
> If this is another specific job for submissions, where can I get the logs from the dockers, please?

Hi

it is the principle of the gating... 😊
we re-install a full fresh ONAP + your patch under review and we test

see https://docs.onap.org/projects/onap-integration/en/latest/onap-integration-ci.html#integration-ci
it has been set since El Alto to be able to give feedback on OOM changes prior to merge...
because before, we had precisely to wait for the merge to observe if we did not "break" everything.

if the gating is OK (no regression on tests => we merge)
if not we do not merge...

The gating is valid on any OOM, SO and pythonsdk-test gerrit changes
it means that if you need to have simulatenously several changes - let's say new docker in OOM + test adaptation required in testsuite it may not work. 
You need to merge on testsuite first (as the testing dockers are regenerated everyday) before submitting (and then automatically gating) on oom

the logs of the dockers are theoretically included in the the test result dashboard => https://logs.onap.org/onap-integration/gating/127471-4/k8s-teardown/k8s-teardown/kubernetes-status/index.html
but here there is no page unfortunately there was a crash durign the test so except the robot test logs, you will have nothing more as the gating labs are continuously reused...
you can ask for a new gating from the patch (any new version of the patch, a change in the commit message or the magic word will retrigger a gate)

as soon as a new patch is submited it is queued and it uses the first free gating labs.
today there are 2 gating labs running on Azure for the whole project.
I am pretty sure there were some video recordings during old DDF on this topic

Patch-set: 4
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>