Update patch set 2
Patch Set 2: -Code-Review

> Patch Set 2:
> 
> @Morgan Richomme Yeah I did see those two failing which was odd but I noticed the same is failing on DCAE's gating run that doesnt have these changes so thinking this could not have caused it. From jira ticket comment "I do see two other smoke tests failing in SO (basic_cnf_macro & basic_network) but they are also failing without these changes so might be an issue in SO, if you see here those two also failed https://gerrit.onap.org/r/c/oom/+/127922 but the difference here is basic_vm passes which I was addressing, thanks" Please keep an eye on the results, as I don't believe this to be the culprit failing those or else it wouldn't have failed outside of this change.

sure, usually we try to compare with the daily that is supposed to be stable.
At the moment it is unfortunately not really the case
basic_network used to be fine on the last 3 daily master runs but a 500 from SO can be a transient error

on the last weekly master https://logs.onap.org/onap-integration/weekly/onap_weekly_pod4_master/2022-02/25_20-26/ the 2 tests were OK, only the basic_vm were affected.
so I assume if it is reproducible and affects other patches, it means that a change breaking master has been merged and your patch is not responsible for that. We need to find when the behavior changed.
Unfortunately I am no more fully involved in ONAP and have no resource to investigate further

Patch-set: 2
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0