[POLICY] Increase hikari connection pool size for PAP & API
We saw some bottlenecks during S3P testing of API component due to the default connection pool size of 10. Setting it to 20 based on the number of parallel threads used for performance tests. The change is required because prior to spring boot migration (Istanbul and earlier), the default connection size was 32 so it was never a problem for running load tests.

Issue-ID: POLICY-4007
Change-Id: Iaaa573d5fce0298e3aff05a27c2e845840a6cb06
Signed-off-by: Prakhar Pandey <prakhar.pandey@bell.ca>